refactor: useTags 파일 분리
